Add some more tests, refine a return type.
Previously, Resttest used an anonymous Kotlin object as a return in a
couple of controller endpoints, which made testing difficult. This was a
temporary prototyping thing that should have been changed before commit,
so now those functions return a simple object.

Resttest did a lot of integration testing with the spring boot
TestRestTemplate, but not a whole lot of unit testing of the endpoints.
This version adds (almost) all of the endpoint functions to unit tests.